Compiles, but I didn,t do the right thing.
The PWM library is not that generic. Previously, I was trying to follow the pwm library in the STM32 hal crate, but that did not work. I believe this PAC is missing some things that that code expects, and I was not able to compile it as my rust is not that good.